fix script

+++ preferences.sh
@@ -3,26 +3,28 @@
 printf \n"======== Applying preferences ========\n"
 
 printf "\t-> guake theme \n"
-	su - dl/guake-colors-solarized/set_dark.sh solarized
+	su $1 -c "dl/guake-colors-solarized/set_dark.sh solarized"
 
 printf "\t-> import .zshrc \n"
 	cp pref/.zshrc /home/$1
 
 printf "\t-> load guake on startup \n"
-	ln -s /usr/share/applications/guake.desktop /etc/xdg/autostart/
+	if [ ! -L /etc/xdg/autostart/guake.desktop ]; then
+		ln -s /usr/share/applications/guake.desktop /etc/xdg/autostart/
+	fi
 
 printf "\t-> copying awesome config \n"
 	if [ ! -d /home/$1/.config/awesome ]; then
 		mkdir /home/$1/.config/awesome
 	fi
-	cp -r dl/awesome-copycats/* /home/$1/.config/awesome
-	cp pref/rc.lua /home/$1/.config/awesome	
+	cp -r dl/awesome-copycats/* /home/$1/.config/awesome/
+	cp pref/rc.lua /home/$1/.config/awesome/
 
 printf "\t-> setting awesome config \n"
-	if [ ! -f /etc/xdg/awesome/rc.lua]; then
+	if [ ! -f /etc/xdg/awesome/rc.lua ]; then
 		rm /etc/xdg/awesome/rc.lua
 	fi
-	ln -s rc.lua /home/$1/.config/awesome/rc.lua
+	ln -s rc.lua /home/$1/.config/awesome/
 	echo "guake&
 
 source /home/$1/.zshrc
